+++ test/kiln_cms/keys/vault_rotation_test.exs
@@ -96,6 +96,7 @@ defmodule KilnCMS.Keys.VaultRotationTest do
       {KilnCMS.CMS.SiteAiProvider, :api_key_encrypted},
       {KilnCMS.CMS.SiteMailRelay, :password_encrypted},
       {KilnCMS.CMS.SiteVapidKey, :private_key_encrypted},
+      {KilnCMS.CMS.StorageProfile, :secret_access_key_encrypted},
       {KilnCMS.CMS.WebhookEndpoint, :secret_encrypted},
       {KilnCMS.Federation.SiteFederation, :private_key_encrypted},
       {KilnCMS.Mail.Settings, :dkim_private_key_encrypted},